backend/requirements_clean.txt
# Flask 核心框架
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1

# 文档处理
python-docx==1.2.0
docxtpl==0.20.1
docxcompose==1.3.6
PyPDF2==3.0.1
openpyxl==3.0.10

# 工具库
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
